Migrate performance, accuracy, and quantization tests to CI registry
Add CI registration to quantization tests:
- quant/test_awq.py (163s) -> stage-b-test-small-1-gpu
- quant/test_marlin_moe.py (200s) -> stage-b-test-small-1-gpu
- test_bnb.py (5s) -> stage-b-test-small-1-gpu
- test_gptqmodel_dynamic.py (102s) -> stage-b-test-small-1-gpu
- test_quantization.py (185s) -> stage-b-test-small-1-gpu
- test_gguf.py (96s) -> stage-b-test-small-1-gpu

Add CI registration to accuracy tests:
- test_eval_accuracy_large.py (300s) -> stage-b-test-small-1-gpu
- test_moe_eval_accuracy_large.py (500s) -> stage-b-test-large-2-gpu

Split performance tests by GPU requirements:
- test_bench_serving_1gpu.py -> stage-b-test-small-1-gpu (5090-compatible)
  - Basic LLM throughput/latency, VLM, LoRA, score/embeddings API
- test_bench_serving_1gpu_large.py -> stage-b-test-large-1-gpu
  - FP8, EAGLE speculative decoding (need H200/SM90+)
- test_bench_serving_2gpu.py -> stage-b-test-large-2-gpu
  - MoE (TP=2), Pipeline Parallel (PP=2)
