+++ test/srt/test_bnb.py
@@ -12,6 +12,7 @@ from types import SimpleNamespace
 import openai
 
 from sglang.srt.utils import kill_process_tree
+from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.run_eval import run_eval
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
@@ -20,6 +21,8 @@ from sglang.test.test_utils import (
     is_in_ci,
     popen_launch_server,
 )
+
+register_cuda_ci(est_time=5, suite="stage-b-test-small-1-gpu")
 
 VISION_MODELS = [
     "unsloth/Qwen2.5-VL-7B-Instruct-bnb-4bit",
