+++ test/srt/test_gptqmodel_dynamic.py
@@ -6,12 +6,15 @@ import torch
 
 from sglang.srt.server_args import set_global_server_args_for_scheduler
 from sglang.srt.utils import kill_process_tree
+from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     DEFAULT_URL_FOR_TEST,
     CustomTestCase,
     popen_launch_server,
 )
+
+register_cuda_ci(est_time=102, suite="stage-b-test-small-1-gpu")
 
 
 def check_quant_method(model_path: str, use_marlin_kernel: bool):
